requirements.txt
requirements:
streamlit
tensorflow
Pillow
numpy

